trying to implemen view on ws

+++ video-service/src/main/java/io/muzoo/scalable/vms/VideoDetailsResponseDTO.java
@@ -30,5 +30,5 @@ public class VideoDetailsResponseDTO {
     @JsonProperty("status")
     private final VideoStatus status;
     @JsonProperty("viewCount")
-    private final int viewCount;
+    private final Long viewCount;
 }
